# remove ISymbol, IArgument, IIdentifierSymbol, IOption and ICommand interfaces