--- a/src/System.CommandLine.Benchmarks/CommandLine/Perf_Parser_NestedCommands.cs
+++ b/src/System.CommandLine.Benchmarks/CommandLine/Perf_Parser_NestedCommands.cs
@@ -55,7 +55,7 @@ namespace System.CommandLine.Benchmarks.CommandLine
             GenerateTestNestedCommands(rootCommand, TestCommandsDepth, TestCommandsDepth);
 
             // Choose only one path from the commands tree for the test arguments string
-            ISymbol currentCmd = rootCommand;
+            Symbol currentCmd = rootCommand;
             while (currentCmd.Children.Count > 0)
             {
                 currentCmd = currentCmd.Children[0];