--- a/src/System.CommandLine/Binding/NoArgumentConversionResult.cs
+++ b/src/System.CommandLine/Binding/NoArgumentConversionResult.cs
@@ -5,7 +5,7 @@ namespace System.CommandLine.Binding
 {
     internal class NoArgumentConversionResult : ArgumentConversionResult
     {
-        internal NoArgumentConversionResult(IArgument argument) : base(argument)
+        internal NoArgumentConversionResult(Argument argument) : base(argument)
         {
         }
     }