--- a/src/System.CommandLine/Binding/TooManyArgumentsConversionResult.cs
+++ b/src/System.CommandLine/Binding/TooManyArgumentsConversionResult.cs
@@ -5,7 +5,7 @@ namespace System.CommandLine.Binding
 {
     internal class TooManyArgumentsConversionResult : FailedArgumentConversionArityResult
     {
-        internal TooManyArgumentsConversionResult(IArgument argument, string errorMessage) : base(argument, errorMessage)
+        internal TooManyArgumentsConversionResult(Argument argument, string errorMessage) : base(argument, errorMessage)
         {
         }
     }