--- a/src/System.CommandLine/CommandLineConfiguration.cs
+++ b/src/System.CommandLine/CommandLineConfiguration.cs
@@ -97,7 +97,7 @@ namespace System.CommandLine
         /// <summary>
         /// Gets the root command.
         /// </summary>
-        public ICommand RootCommand { get; }
+        public Command RootCommand { get; }
 
         internal ResponseFileHandling ResponseFileHandling { get; }