… src/System.CommandLine.ApiCompatibility.Tests/ApiCompatibilityApprovalTests.System_CommandLine_api_is_not_changed.approved.txt
… src/System.CommandLine/IArgument.cs
… src/System.CommandLine/ICommand.cs
… src/System.CommandLine/IIdentifierSymbol.cs
… src/System.CommandLine/IOption.cs
… src/System.CommandLine/ISymbol.cs
… src/System.CommandLine/Option.cs
… src/System.CommandLine/Parsing/ArgumentResult.cs
… src/System.CommandLine/Parsing/CommandArgumentNode.cs
… src/System.CommandLine/Parsing/CommandResult.cs
… src/System.CommandLine/Parsing/OptionArgumentNode.cs
… src/System.CommandLine/Parsing/OptionNode.cs
… src/System.CommandLine/Parsing/OptionResult.cs
… src/System.CommandLine/Parsing/ParseOperation.cs
… src/System.CommandLine/Parsing/ParseResult.cs
… src/System.CommandLine/Parsing/ParseResultExtensions.cs
… src/System.CommandLine/Parsing/ParseResultVisitor.cs
… src/System.CommandLine/Parsing/RootCommandNode.cs
… src/System.CommandLine/Parsing/RootCommandResult.cs
… src/System.CommandLine/Parsing/StringExtensions.cs
… src/System.CommandLine/Parsing/SymbolResult.cs
… src/System.CommandLine/Parsing/SymbolResultExtensions.cs
… src/System.CommandLine/Parsing/Token.cs
… src/System.CommandLine/Symbol.cs
… src/System.CommandLine/SymbolExtensions.cs